added file structure
