+++ SortingHat/Models/Learner.swift
@@ -1,0 +1,14 @@
+//
+//  Learner.swift
+//  SortingHat
+//
+//  Created by Nicholas Melekian on 8/1/23.
+//
+
+import Foundation
+
+struct Learner: Identifiable {
+    let id = UUID()
+    var first: String
+    var last: String
+}
