+++ SortingHat/Models/Team.swift
@@ -1,0 +1,12 @@
+//
+//  Team.swift
+//  SortingHat
+//
+//  Created by Nicholas Melekian on 8/1/23.
+//
+
+import Foundation
+
+struct Team {
+    var members: [Learner]
+}
